requirements.txt
chromedriver-binary==88.0.4324.96.0
discord.py==1.6.0
selenium==3.141.0
tweepy==3.10.0
